Implement automatic SDK version detection
Implement automatic SDK detection based on the SDK_VERSION file in the
upstream Zephyr repository. Lookup is in the "zephyr" module directory
if available, "zephyr" directory if not.

Switch the default SDK to "auto" so that the workflow will automatically
pick the current one for the Zephyr checkout specified in the manifest
file.

Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>