Add a base-path action input variable
Add a base-path action input variable and use it on all steps that are
using paths, annotate that app-path can only be a single directory. The
west tool creates the workspace on the parent of app-path so this is
inline with that behavior, and allows using the action on repositories
where the manifest lives in a subdirectory.

Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>