action: encode the repo name in the SDK cache key
The Zephyr SDK install has a step where it tweaks the library paths of
the host utilities so match the full install path and have them run
without having to set LD_LIBRARY_PATH. In the current setup that path
includes the repository name, which means that if the repository is
renamed the cached SDK is going to end up having a non working path
leading to a confusing error, typically dtc not found.

Make the repository name part of the SDK cache key, so that if the
repository is renamed, the cached sdk with the wrong path is not reused.

Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>