action: use $HOME/zephyr-sdk for the Zephyr SDK
The Zephyr SDK install has a step where it tweaks the library paths of
the host utilities so match the full install path and have them run
without having to set LD_LIBRARY_PATH. In the current setup that path
includes the repository name, which means that if the repository is
renamed the cached SDK is going to end up having a non working path
leading to a confusing error, typically dtc not found.

Move the SDK install to the home directory, so that the path does not
depend on the repository name. The SDK is found using a CMake package so
it can really be installed anywhere.

Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>