action: pin GitHub actions via SHA
There were several issues with CI actions and moved tags on GitHub in
the past, therefore a lot of repositories activated the function to
enforce fulllength sha pins for actions.

This also applies recursively, and this action does not pin used
actions itself, making it not usable in a lot of other repos.

To improve adoption and make the action also more deterministic, we pin
the versions of used actions in this commit.

Signed-off-by: Stefan Kraus <dev@stefankraus.org>