action: do not re-install ninja
After many years asking, Github runners finally got ninja by
default in April 2025:
- https://github.com/actions/runner-images/issues/11235

So, stop re-installing ninja - at least not unless there is a specific
and documented reason to do so. This saves a little bit of time and
fixes the following warning in hello_world_multiplatform.yaml:

https://github.com/zephyrproject-rtos/zephyr/actions/runs/19853412716
```
build (macos-14)
ninja 1.13.2 is already installed and up-to-date.
To reinstall 1.13.2, run: brew reinstall ninja
```

Do not change the Windows build yet because the build environment is
always much more complicated on Windows, so let's first make sure this
is fine on Linux and macOS.

Signed-off-by: Marc Herbert <marc.herbert@intel.com>